fix(rust): metrics percentiles as f64 (wire parity with Node number|null)

+++ rust/protocol/src/player.rs
@@ -187,9 +187,9 @@ pub struct MetricsSubscribe {
 #[serde(rename_all = "camelCase")]
 pub struct MetricPercentiles {
     #[ts(optional)]
-    pub p50: Option<i32>,
+    pub p50: Option<f64>,
     #[ts(optional)]
-    pub p95: Option<i32>,
+    pub p95: Option<f64>,
     pub count: i32,
 }
 
@@ -276,13 +276,13 @@ mod tests {
     fn test_metrics_health_snapshot_roundtrip() {
         let snapshot = MetricsHealthSnapshot {
             rtt: MetricPercentiles {
-                p50: Some(50),
-                p95: Some(100),
+                p50: Some(50.0),
+                p95: Some(100.0),
                 count: 10,
             },
             clock_offset: MetricPercentiles {
-                p50: Some(5),
-                p95: Some(15),
+                p50: Some(5.0),
+                p95: Some(15.0),
                 count: 10,
             },
             answer_ack: MetricPercentiles {
